requirements.txt
discord.py
python-dotenv~=1.0.1
